requirements.txt
django==4.1.3
mysqlclient==2.1.1
python-dotenv==0.21.0
djangorestframework==3.14.0
markdown
django-filter==22.1
drf-nested-routers==0.93.4
pre-commit==2.20.0
openpyxl==3.0.10
gunicorn==20.1.0
elasticsearch==6.4.0